requirements.txt
asgiref==3.9.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cloudinary==1.44.1
crispy-bootstrap4==2025.6
cryptography==46.0.4
dj-database-url==3.1.1
Django==5.2.5
django-allauth==65.14.0
django-autoslug==1.9.9
django-ckeditor==6.7.3
django-ckeditor-5==0.2.19
django-cloudinary-storage==0.3.0
django-crispy-forms==2.5
django-js-asset==3.1.2
django-tinymce==5.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.0.3
idna==3.11
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.6.3
whitenoise==6.11.0
